# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Example Golang REST API",
    "title": "Go Example REST API",
    "contact": {
      "name": "AJackTi",
      "url": "https://github.com/AJackTi",
      "email": "dtrong97vn@gmail.com"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/auth/all": {
      "get": {
        "description": "Get the list of all users",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Get users",
        "parameters": [
          {
            "type": "integer",
            "format": "page",
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "size",
            "description": "number of elements per page",
            "name": "size",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "orderBy",
            "description": "filter name",
            "name": "orderBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UsersList"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
    "/auth/find": {
      "get": {
        "description": "Find user by name",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Find by name",
        "parameters": [
          {
            "type": "string",
            "format": "username",
            "description": "username",
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UsersList"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "description": "Get current user by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Get user by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
...
    "/auth/token": {
      "get": {
        "description": "Get CSRF token, required auth session cookie",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Get CSRF token",
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
    "/auth/{id}": {
      "get": {
        "description": "get string by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "get user by id",
        "parameters": [
          {
            "type": "integer",
            "description": "user_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      },
      "put": {
        "description": "update existing user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Update user",
        "parameters": [
          {
            "type": "integer",
            "description": "user_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          }
        }
      },
      "delete": {
        "description": "some description",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Delete user account",
        "parameters": [
          {
            "type": "integer",
            "description": "user_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
    "/auth/{id}/avatar": {
      "post": {
        "description": "Post user avatar image",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Post avatar",
        "parameters": [
          {
            "type": "file",
            "description": "Body with image file",
            "name": "file",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "format": "bucket",
            "description": "aws s3 bucket",
            "name": "bucket",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "user_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "httpErrors.RestError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      }
    },
...
    "models.User": {
      "type": "object",
      "required": ["first_name", "last_name", "password"],
      "properties": {
        "about": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "login_date": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "postcode": {
          "type": "integer"
        },
        "role": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        }
      }
    },
    "models.UsersList": {
      "type": "object",
      "properties": {
        "has_more": {
          "type": "boolean"
        },
        "page": {
          "type": "integer"
        },
        "size": {
          "type": "integer"
        },
        "total_count": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.User"
          }
        }
      }
    }
  }
}